Clean code: Rename BasicItineraryTestCase to BasicPathTestCase.
