+++ src/test/java/org/opentripplanner/transit/raptor/rangeraptor/path/PathMapperTest.java
@@ -1,7 +1,25 @@
 package org.opentripplanner.transit.raptor.rangeraptor.path;
 
+import static org.junit.Assert.assertEquals;
+import static org.opentripplanner.transit.raptor._data.stoparrival.BasicPathTestCase.ACCESS_COST;
+import static org.opentripplanner.transit.raptor._data.stoparrival.BasicPathTestCase.ACCESS_START;
+import static org.opentripplanner.transit.raptor._data.stoparrival.BasicPathTestCase.BASIC_PATH_AS_STRING;
+import static org.opentripplanner.transit.raptor._data.stoparrival.BasicPathTestCase.EGRESS_COST;
+import static org.opentripplanner.transit.raptor._data.stoparrival.BasicPathTestCase.EGRESS_END;
+import static org.opentripplanner.transit.raptor._data.stoparrival.BasicPathTestCase.LINE_11_COST;
+import static org.opentripplanner.transit.raptor._data.stoparrival.BasicPathTestCase.LINE_21_COST;
+import static org.opentripplanner.transit.raptor._data.stoparrival.BasicPathTestCase.LINE_31_COST;
+import static org.opentripplanner.transit.raptor._data.stoparrival.BasicPathTestCase.SLACK_PROVIDER;
+import static org.opentripplanner.transit.raptor._data.stoparrival.BasicPathTestCase.TOTAL_COST;
+import static org.opentripplanner.transit.raptor._data.stoparrival.BasicPathTestCase.TRIP_DURATION;
+import static org.opentripplanner.transit.raptor._data.stoparrival.BasicPathTestCase.TX_COST;
+import static org.opentripplanner.transit.raptor._data.stoparrival.BasicPathTestCase.basicTripByReverseSearch;
+import static org.opentripplanner.transit.raptor._data.stoparrival.BasicPathTestCase.lifeCycle;
+import static org.opentripplanner.transit.raptor._data.transit.TestTransfer.walk;
+import static org.opentripplanner.transit.raptor.api.transit.RaptorCostConverter.toOtpDomainCost;
+
 import org.junit.Test;
-import org.opentripplanner.transit.raptor._data.stoparrival.BasicItineraryTestCase;
+import org.opentripplanner.transit.raptor._data.stoparrival.BasicPathTestCase;
 import org.opentripplanner.transit.raptor._data.stoparrival.Egress;
 import org.opentripplanner.transit.raptor._data.transit.TestTripSchedule;
 import org.opentripplanner.transit.raptor.api.path.Path;
@@ -9,30 +27,12 @@ import org.opentripplanner.transit.raptor.api.path.PathLeg;
 import org.opentripplanner.transit.raptor.rangeraptor.WorkerLifeCycle;
 import org.opentripplanner.util.time.TimeUtils;
 
-import static org.junit.Assert.assertEquals;
-import static org.opentripplanner.transit.raptor._data.stoparrival.BasicItineraryTestCase.ACCESS_COST;
-import static org.opentripplanner.transit.raptor._data.stoparrival.BasicItineraryTestCase.ACCESS_START;
-import static org.opentripplanner.transit.raptor._data.stoparrival.BasicItineraryTestCase.BASIC_PATH_AS_STRING;
-import static org.opentripplanner.transit.raptor._data.stoparrival.BasicItineraryTestCase.EGRESS_COST;
-import static org.opentripplanner.transit.raptor._data.stoparrival.BasicItineraryTestCase.EGRESS_END;
-import static org.opentripplanner.transit.raptor._data.stoparrival.BasicItineraryTestCase.LINE_11_COST;
-import static org.opentripplanner.transit.raptor._data.stoparrival.BasicItineraryTestCase.LINE_21_COST;
-import static org.opentripplanner.transit.raptor._data.stoparrival.BasicItineraryTestCase.LINE_31_COST;
-import static org.opentripplanner.transit.raptor._data.stoparrival.BasicItineraryTestCase.SLACK_PROVIDER;
-import static org.opentripplanner.transit.raptor._data.stoparrival.BasicItineraryTestCase.TOTAL_COST;
-import static org.opentripplanner.transit.raptor._data.stoparrival.BasicItineraryTestCase.TRIP_DURATION;
-import static org.opentripplanner.transit.raptor._data.stoparrival.BasicItineraryTestCase.TX_COST;
-import static org.opentripplanner.transit.raptor._data.stoparrival.BasicItineraryTestCase.basicTripByReverseSearch;
-import static org.opentripplanner.transit.raptor._data.stoparrival.BasicItineraryTestCase.lifeCycle;
-import static org.opentripplanner.transit.raptor._data.transit.TestTransfer.walk;
-import static org.opentripplanner.transit.raptor.api.transit.RaptorCostConverter.toOtpDomainCost;
-
 public class PathMapperTest {
 
     @Test
     public void mapToPathForwardSearch() {
         // Given:
-        Egress egress = BasicItineraryTestCase.basicTripByForwardSearch();
+        Egress egress = BasicPathTestCase.basicTripByForwardSearch();
         DestinationArrival<TestTripSchedule> destArrival = new DestinationArrival<>(
                 walk(egress.previous().stop(), egress.durationInSeconds()),
                 egress.previous(),
@@ -40,7 +40,7 @@ public class PathMapperTest {
                 egress.additionalCost()
         );
 
-        WorkerLifeCycle lifeCycle = BasicItineraryTestCase.lifeCycle();
+        WorkerLifeCycle lifeCycle = BasicPathTestCase.lifeCycle();
         PathMapper<TestTripSchedule> mapper = new ForwardPathMapper<>(
             SLACK_PROVIDER,
             lifeCycle
